requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
psycopg2-binary
flask-migrate
python-dotenv
